add set group manager

+++ client/src/js/component/group/member.js
@@ -1,6 +1,10 @@
 import React from 'react';
+import ReactDOM from 'react-dom';
 import request from '../../utils/request';
 import DataTable from '../dataTable/dataTable';
+import { browserHistory } from 'react-router';
+import { toast } from 'react-toastify';
+import { ModalManager, Icon } from '../common';
 
 class Member extends React.Component {
   constructor(props) {
@@ -9,6 +13,7 @@ class Member extends React.Component {
     this.state = {
       userName: '',
     };
+    this.user = JSON.parse(localStorage.getItem('user'));
   }
 
   handleSearch = (searchWord) => {
@@ -18,45 +23,92 @@ class Member extends React.Component {
   }
 
   removeMember = () => {
-    request({
-      url: '/group',
-      method: 'delete',
-      data: {
-        ids: this.ids,
-        groupId: this.props.group.id,
-        groupName: this.props.group.group_name,
-      },
-    }).then(resp => {
-      if (resp.code === '1') {
-        this.table.refresh();
+    ModalManager.confirm({
+      content: '确定要移除?',
+      onOk: async () => {
+        await request({
+          url: '/group',
+          method: 'delete',
+          data: {
+            ids: this.ids,
+            groupId: this.props.group.id,
+            groupName: this.props.group.group_name,
+          },
+        }).then(resp => {
+          if (resp.code === '1') {
+            this.table.refresh();
+            toast.success('移除成功', toastOption);
+            socket.emit('updateMyGroupList', this.user.user_id);
+            this.handleRedirect(this.user.user_id);
+            return true;
+          }
+        });
       }
-    });
+    })
   }
 
   handleCheckedChange = (ids) => {
     this.ids = ids;
   }
 
+  convertParms = (data) => {
+    return {
+      retList: data.retList[0].ret,
+      total: data.retList[0].count,
+    };
+  }
+
+  setManager = (userId) => {
+    request({
+      url: '/manager',
+      method: 'post',
+      data: {
+        userId,
+        groupId: this.props.group.id,
+      },
+    }).then(resp => {
+      if (resp.code === '1') {
+        this.table.refresh();
+        toast.success('设置管理员成功', toastOption);
+      }
+    });
+  }
+
+  handleRedirect = (userId) => {
+    const modalContainer = document.getElementsByClassName('modal-container')[0];
+    const unmountResult = ReactDOM.unmountComponentAtNode(modalContainer);
+    if (unmountResult && modalContainer.parentNode) {
+      modalContainer.parentNode.removeChild(modalContainer);
+    }
+    browserHistory.push(`/chat/${userId}`);
+  }
+
   render() {
     const { count } = this.props.countInfo;
     const pageSize = 1;
+    const { isCreater, isManager, createrId } = this.props;
     const columns = [
       { label: '姓名', field: 'username' },
       { label: '年龄', field: 'age' },
       { label: '性别', field: 'sex' },
       { label: '手机号', field: 'phone' },
-      {label: '邮箱', field: 'email'},
+      { label: '邮箱', field: 'email' },
+      { label: '操作', template: row => (
+        <div>
+          {isCreater && <Icon name="user" className="tooltip" data-tooltip="设置管理员" onClick={this.setManager.bind(this, row.id)}/>}
+          <Icon name="comment-alt" prefix1="far" className="tooltip" data-tooltip="发消息" onClick={this.handleRedirect.bind(this, row.id)}/>
+        </div>
+      )},
     ];
     const params = {
       url: '/group',
       data: {
         groupId: this.props.group.id,
-        type: '4',
-        userName: this.state.userName,
+        userId: createrId,
+        type: '8',
         limit: pageSize,
       },
     };
-
     return (
       <DataTable
         rowKey='id'
@@ -64,13 +116,17 @@ class Member extends React.Component {
         params={params}
         columns={columns}
         showToolbar={true}
+        checkable={(row) => row.id == createrId}
+        convertParms={this.convertParms}
         toolbar={{
           hasSearch: true,
           hasRefresh: false,
           onSearch: this.handleSearch,
           items: [{
             align: 'left',
-            content: <button className="button is-danger" onClick={this.removeMember}>移除</button>,
+            content: <div>
+              {(isCreater || isManager) && <button className="button is-danger is-small" onClick={this.removeMember}>移除</button>}
+            </div>,
           }]
         }}
         noDataText={t('No Data')}
